Move developer-related notifications to Developers
