+++ db/seeds/notifications.rb
@@ -8,15 +8,15 @@ Message.find_or_create_by!(conversation:, sender: developer, body: "Let's chat o
 Notification.find_or_create_by!(type: NewMessageNotification.name, recipient: developer.user, params: {message:, conversation:})
 
 # DeveloperMailer#welcome
-Notification.find_or_create_by!(type: WelcomeDeveloperNotification.name, recipient: developer.user, params: {developer:})
+Notification.find_or_create_by!(type: Developers::WelcomeNotification.name, recipient: developer.user, params: {developer:})
 
 # DeveloperMailer#invisiblize
 invisible_developer = User.find_by(email: "invisible@example.com").developer
-Notification.find_or_create_by!(type: InvisiblizeDeveloperNotification.name, recipient: invisible_developer.user, params: {developer: invisible_developer})
+Notification.find_or_create_by!(type: Developers::InvisiblizeNotification.name, recipient: invisible_developer.user, params: {developer: invisible_developer})
 
 # DeveloperMailer#stale
 stale_developer = User.find_by(email: "stale@example.com").developer
-Notification.find_or_create_by!(type: StaleDeveloperNotification.name, recipient: stale_developer.user, params: {developer: stale_developer})
+Notification.find_or_create_by!(type: Developers::ProfileReminderNotification.name, recipient: stale_developer.user, params: {developer: stale_developer})
 
 # AdminMailer#new_business
 admin = User.find_by(email: "admin@example.com")
