+++ test/mailers/developer_mailer_test.rb
@@ -2,15 +2,15 @@ require "test_helper"
 
 class DeveloperMailerTest < ActionMailer::TestCase
   test "invisiblize" do
-    InvisiblizeDeveloperNotification.with(developer:).deliver(recipient)
+    Developers::InvisiblizeNotification.with(developer:).deliver(recipient)
   end
 
   test "stale" do
-    StaleDeveloperNotification.with(developer:).deliver(recipient)
+    Developers::ProfileReminderNotification.with(developer:).deliver(recipient)
   end
 
   test "welcome" do
-    WelcomeDeveloperNotification.with(developer:).deliver(recipient)
+    Developers::WelcomeNotification.with(developer:).deliver(recipient)
   end
 
   def developer
