+++ test/mailers/previews/developer_mailer_preview.rb
@@ -1,16 +1,16 @@
 class DeveloperMailerPreview < ActionMailer::Preview
   def invisiblize
-    notification = Notification.where(type: InvisiblizeDeveloperNotification.to_s).first
+    notification = Notification.where(type: Developers::InvisiblizeNotification.to_s).first
     DeveloperMailer.with(record: notification, recipient: notification.recipient).invisiblize
   end
 
   def stale
-    notification = Notification.where(type: StaleDeveloperNotification.to_s).first
+    notification = Notification.where(type: Developers::ProfileReminderNotification.to_s).first
     DeveloperMailer.with(record: notification, recipient: notification.recipient).stale
   end
 
   def welcome
-    notification = Notification.where(type: WelcomeDeveloperNotification.to_s).first
+    notification = Notification.where(type: Developers::WelcomeNotification.to_s).first
     DeveloperMailer.with(record: notification, recipient: notification.recipient).welcome
   end
 end
